Magic item specification changes
change: The enchantment suitability of armor affects the effect of magic.
change: Changed the effects of some red and green magic items.
change: When using a hoe on a leaf block of HaC trees, a sapling will drop.
fix: Fixed the block destruction process of the lumberjack effect.
fix: Fixed the behavior of the scythe.
fix: Fixed the harvesting items of Camellia and Schima.